Moves controller classes into /controllers folder
