+++ src/controllers/game_state.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class GameState(Enum):
+  INTRO  = 0
+  TRAVEL = 1
+  BATTLE = 2
